Split scripts for jinja differentiation in copier.yml
