+++ template/package.json.jinja
@@ -81,6 +81,7 @@
         "dev:django": "python manage.py runserver_plus"
         {% endif %}
         "build": "webpack --config webpack/prod.config.js",
+        {% endif %}
     {% else %}
         "dev": "",
         "build": ""
